Extract about/help/shortcuts window into WindowHelpMixin
Continue decomposing window.py: move show_about_dialog, open_help_url,
show_shortcuts_window and its builder helpers (_build_shortcuts_window,
_add_safe_current_shortcuts, _get_safe_current_shortcuts,
_add_fallback_shortcuts) verbatim into sshpilot/window_help.py as
WindowHelpMixin, inherited by MainWindow. window.py shrinks ~318 lines.
Pure code move — signatures and self. state access unchanged.

Add tests/test_window_help_mixin.py asserting each method resolves to the
window_help module (catches a stray body copy re-shadowing the mixin).
